Collapse errors dir into lib
