+++ packages/viewer/src/lib/errors.ts
@@ -1,4 +1,13 @@
-import { CustomError } from './customError';
+export class CustomError extends Error {
+  public constructor(message: string, e?: Error) {
+    super();
+
+    this.message =
+      e != null ? `${message} Nested exception is: ${e.message}` : message;
+    this.stack = e?.stack;
+    this.name = this.constructor.name;
+  }
+}
 
 export class ExpiredCredentialsError extends CustomError {}
