--- packages/viewer/src/lib/errors/index.ts
@@ -1,1 +1,0 @@
-export * from './errors';
